refactor(edgeless): remove tool action callback, every tool management by itself
